+++ fabric/src/main/java/com/lx862/mtrscripting/mod/MTRScripting.java
@@ -1,8 +1,11 @@
 package com.lx862.mtrscripting.mod;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.mtr.mapping.holder.Identifier;
 
 public class MTRScripting {
+    public static final Logger LOGGER = LogManager.getLogger("MTR Scripting via JCM");
     private static final String MOD_ID = "mtrscripting";
 
     public static Identifier id(String path) {