+++ fabric/src/main/java/com/lx862/mtrscripting/util/BackgroundWorkerJS.java
@@ -1,6 +1,6 @@
 package com.lx862.mtrscripting.util;
 
-import com.lx862.mtrscripting.ScriptManager;
+import com.lx862.mtrscripting.mod.MTRScripting;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -13,7 +13,7 @@ public class BackgroundWorkerJS {
             try {
                 scriptTask.run();
             } catch (Exception e) {
-                ScriptManager.LOGGER.error("Script error while running background task!", e);
+                MTRScripting.LOGGER.error("Script error while running background task!", e);
             }
         });
     }